+++ R/trakt.user.friendfollow.R
@@ -4,6 +4,8 @@
 #' of both following and being followed by a user.
 #' Since no OAuth2 methods are supported yet, the specified user mustn't be private.
 #' @param user Target user. Defaults to \code{getOption("trakt.username")}
+#' @param extended Either \code{min} for standard info, \code{full} for details or \code{full,images}
+#' for additional avatar URLs.
 #' @return A \code{data.frame} containing user information.
 #' @export
 #' @note See \href{http://docs.trakt.apiary.io/reference/users/friends/get-friends}{the trakt API docs for further info}
@@ -13,13 +15,13 @@
 #' get_trakt_credentials() # Set required API data/headers
 #' trakt.user.friends("jemus42")
 #' }
-trakt.user.friends <- function(user = getOption("trakt.username")){
+trakt.user.friends <- function(user = getOption("trakt.username", extended = "min")){
   if (is.null(user) && is.null(getOption("trakt.username"))){
     stop("No username is set.")
   }
 
   # Construct URL, make API call
-  url      <- build_trakt_url("users", user, "friends")
+  url      <- build_trakt_url("users", user, "friends", extended = extended)
   response <- trakt.api.call(url = url)
 
   if (identical(response, list())){
@@ -36,6 +38,8 @@ trakt.user.friends <- function(user = getOption("trakt.username")){
 #' \code{trakt.user.followers} pulls a user's followers
 #' Since no OAuth2 methods are supported yet, the specified user mustn't be private.
 #' @param user Target user. Defaults to \code{getOption("trakt.username")}
+#' @param extended Either \code{min} for standard info, \code{full} for details or \code{full,images}
+#' for additional avatar URLs.
 #' @return A \code{data.frame} containing user information.
 #' @export
 #' @note See \href{http://docs.trakt.apiary.io/reference/users/followers/get-followers}{the trakt API docs for further info}
@@ -45,13 +49,13 @@ trakt.user.friends <- function(user = getOption("trakt.username")){
 #' get_trakt_credentials() # Set required API data/headers
 #' trakt.user.followers("jemus42")
 #' }
-trakt.user.followers <- function(user = getOption("trakt.username")){
+trakt.user.followers <- function(user = getOption("trakt.username", extended = "min")){
   if (is.null(user) && is.null(getOption("trakt.username"))){
     stop("No username is set.")
   }
 
   # Construct URL, make API call
-  url      <- build_trakt_url("users", user, "followers")
+  url      <- build_trakt_url("users", user, "followers", extended = extended)
   response <- trakt.api.call(url = url)
 
   if (identical(response, list())){
@@ -68,6 +72,8 @@ trakt.user.followers <- function(user = getOption("trakt.username")){
 #' \code{trakt.user.following} pulls a user's followings.
 #' Since no OAuth2 methods are supported yet, the specified user mustn't be private.
 #' @param user Target user. Defaults to \code{getOption("trakt.username")}
+#' @param extended Either \code{min} for standard info, \code{full} for details or \code{full,images}
+#' for additional avatar URLs.
 #' @return A \code{data.frame} containing user information.
 #' @export
 #' @note See \href{http://docs.trakt.apiary.io/reference/users/following/get-following}{the trakt API docs for further info}
@@ -77,13 +83,13 @@ trakt.user.followers <- function(user = getOption("trakt.username")){
 #' get_trakt_credentials() # Set required API data/headers
 #' trakt.user.following("jemus42")
 #' }
-trakt.user.following <- function(user = getOption("trakt.username")){
+trakt.user.following <- function(user = getOption("trakt.username", extended = "min")){
   if (is.null(user) && is.null(getOption("trakt.username"))){
     stop("No username is set.")
   }
 
   # Construct URL, make API call
-  url      <- build_trakt_url("users", user, "following")
+  url      <- build_trakt_url("users", user, "following", extended = extended)
   response <- trakt.api.call(url = url)
 
   if (identical(response, list())){